Move the login page to TS and functional components with hooks.
