+++ client/app.tsx
@@ -5,7 +5,7 @@ import { isLoggedIn } from './auth/auth-utils'
 import { EmailVerificationUi } from './auth/email-verification'
 import { ForgotPassword, ForgotUser, ResetPassword } from './auth/forgot'
 import { LoggedInFilter } from './auth/logged-in-filter'
-import Login from './auth/login'
+import { Login } from './auth/login'
 import { Signup } from './auth/signup'
 import { ConnectedDialogOverlay } from './dialogs/connected-dialog-overlay'
 import { usePixelShover } from './dom/pixel-shover'
